Unify healtcheck. Remove extra call to web3.
