+++ blockchains/matic/matic_worker.js
@@ -59,22 +59,6 @@ class MaticWorker extends BaseWorker {
   async init() {
     this.lastConfirmedBlock = await this.web3.eth.getBlockNumber() - constants.CONFIRMATIONS
   }
-
-  healthcheckExportTimeout() {
-    const timeFromLastExport = Date.now() - this.lastExportTime
-    const isExportTimeoutExceeded = timeFromLastExport > constants.EXPORT_TIMEOUT_MLS
-    if (isExportTimeoutExceeded) {
-      return Promise.reject(`Time from the last export ${timeFromLastExport}ms exceeded limit  ${constants.EXPORT_TIMEOUT_MLS}ms.`)
-    } else {
-      return Promise.resolve()
-    }
-  }
-
-  healthcheck() {
-    return this.web3.eth.getBlockNumber()
-    .then(this.healthcheckExportTimeout())
-  }
-
 }
 
 module.exports = {
